fix : 해당 열에 입력된 값이 범위에서 벗어나므로 orderNumber 데이터 타입 변경

+++ src/main/java/com/dalrun/dto/OrderDto.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 
 public class OrderDto implements Serializable{
 	
-    private int orderNumber;
+    private long orderNumber;
     private String memId;
     private String orderName;
     private String orderAddress;
@@ -40,11 +40,11 @@ public class OrderDto implements Serializable{
         this.productId = productId;
     }
 
-    public int getOrderNumber() {
+    public long getOrderNumber() {
         return orderNumber;
     }
 
-    public void setOrderNumber(int orderNumber) {
+    public void setOrderNumber(long orderNumber) {
         this.orderNumber = orderNumber;
     }
